+++ test/index.js
@@ -3,4 +3,5 @@
 describe('Falco', () => {
   require('./umd')
   require('./config')
+  require('./dir')
 })